requirements.txt
mediapipe 
opencv-python 
numpy
openai
azure.cognitiveservices.speech
python-dotenv
httpx